#37 Implement merged timer functions
