+++ SmartCarMQTTApp/app/src/main/AndroidManifest.xml
@@ -15,9 +15,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SmartCarMQTTApp">
         <activity
-            android:name=".TimerActivity"
-            android:exported="false" />
-        <activity
             android:name=".QuizQuestionActivity"
             android:exported="false" />
         <activity
